Update GetRequiredRouteConfig to read from feature
